[#1638] - Extrae mapImagery a un archivo dedicado de _utils
mapImagery (+ hashString/seededShuffle) se mueve de functions.ts a
_utils/storylist-imagery.functions.ts para no engrosar functions.ts, siguiendo el patron
de media-sources.functions.ts. La importan el repo de storylist y el mapper de la landing
(mapStorylistTeasers). Sus tests se mueven a storylist-imagery.functions.spec.ts.
